Zip docset before uploading artifact
This is a temporary workaround for the fact that filenames with colons
are written out, which causes an error because colons are not supported
in paths on Windows. One solution would be to just not include the onnx
docs in the output, since this is the only place the filenames contain
the colons.